new(LoginViewModel): new view model with biometric autheticacion added
